Add autogenerated data and routines for AdvancedExtensions module in ProjectC
This commit introduces a new file, data_projectC.py, which contains data and routines to test the AdvancedExtensions module. The file is automatically generated and should not be modified manually. The author of this file is François Lauze from the University of Copenhagen, dated April 14, 2024. The random state used for generation is also included.